+++ src/lib/Switch.svelte
@@ -93,5 +93,5 @@
 </div>
 
 <style lang="scss">
-	@use './assets/main.scss' as *;
+	@use './assets/switch' as *;
 </style>